+++ OtterEngine/Entity/Plane.cpp
@@ -35,7 +35,7 @@ Plane::Plane(const Graphics& graphics, const Vector3& translation, const Vector3
 
 		// buffers
 		s_commonResources.push_back(std::make_unique<VertexBuffer>(graphics, 
-			s_vertices.data(), sizeof(VertexTexture), s_vertices.size()));
+			s_vertices.data(), static_cast<unsigned int>(sizeof(VertexTexture)), s_vertices.size()));
 		s_commonResources.push_back(std::make_unique<IndexBuffer>(graphics, s_indices));
 	}
